+++ node/middlewares/events/compareOrder.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import {
-  CustomError,
   convertGlovoProductsToCompare,
   isValidAffiliateId,
-} from '../utils'
+  ServiceError,
+} from '../../utils'
 
 export async function compareOrder(
   ctx: StatusChangeContext,
@@ -34,11 +34,7 @@ export async function compareOrder(
     const orderRecord = await recordsManager.getOrderRecord(orderId)
 
     if (!orderRecord) {
-      throw new CustomError({
-        message: `The record for the order ${orderId} was not found`,
-        status: 500,
-        payload: { stores, affiliateId: orderIdAffiliate },
-      })
+      throw new Error(`The record for the order ${orderId} was not found`)
     }
 
     // check if order has changed
@@ -58,9 +54,11 @@ export async function compareOrder(
       if (comparison[invoicedItem.id]) {
         // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
         comparison[invoicedItem.id].quantity += invoicedItem.quantity
-      } else {
-        comparison[invoicedItem.id] = invoicedItem
+
+        continue
       }
+
+      comparison[invoicedItem.id] = invoicedItem
     }
 
     for (const receivedItem of receivedItems) {
@@ -127,11 +125,12 @@ export async function compareOrder(
 
     await next()
   } catch (error) {
-    throw new CustomError({
-      message: `Order comparison for order ${orderId} failed`,
-      status: 500,
-      payload: error,
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason: error.reason ?? `Order comparison for order ${orderId} failed`,
+      metric: 'orders',
+      data: { body },
+      error: error.response?.data,
     })
   }
 }