+++ node/middlewares/events/updateGlovoOrderStatus.ts
@@ -1,9 +1,9 @@
 import {
-  CustomError,
   getStoreInfoFromStoreId,
   isValidAffiliateId,
+  ServiceError,
   setGlovoStatus,
-} from '../utils'
+} from '../../utils'
 
 export async function updateGlovoOrderStatus(ctx: StatusChangeContext) {
   const {
@@ -57,11 +57,16 @@ export async function updateGlovoOrderStatus(ctx: StatusChangeContext) {
       glovoPayload,
     })
   } catch (error) {
-    throw new CustomError({
-      message: `Glovo order ${glovoPayload.glovoOrderId} status update failed`,
-      status: 500,
-      payload: error,
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason:
+        error.reason ??
+        `Glovo order ${glovoPayload.glovoOrderId} status update failed`,
+      metric: 'orders',
+      data: {
+        body,
+      },
+      error: error.response?.data,
     })
   }
 }