+++ node/middlewares/routes/getGlovoMenu.ts
@@ -1,4 +1,4 @@
-import { CustomError } from '../utils'
+import { ServiceError } from '../../utils'
 
 export async function getGlovoMenu(ctx: Context) {
   const {
@@ -10,11 +10,11 @@ export async function getGlovoMenu(ctx: Context) {
 
     ctx.body = glovoMenu
   } catch (error) {
-    throw new CustomError({
-      message: `There was a problem getting the Glovo menu`,
-      status: 500,
-      payload: error,
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason: `There was a problem getting the Glovo menu`,
+      metric: 'menu',
+      error: error.response?.data,
     })
   }
 }