+++ node/middlewares/routes/saveOrderRecord.ts
@@ -1,4 +1,4 @@
-import { CustomError } from '../utils'
+import { ServiceError } from '../../utils'
 
 export async function saveOrderRecord(ctx: Context) {
   const {
@@ -21,11 +21,12 @@ export async function saveOrderRecord(ctx: Context) {
 
     ctx.status = 201
   } catch (error) {
-    throw new CustomError({
-      message: `Unable to save order record for order ${vtexOrder}`,
-      status: 500,
-      payload: { orderId, orderRecord: data },
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason: `Unable to save order record for order ${vtexOrder}`,
+      metric: 'orders',
+      data: { orderId, orderRecord: data },
+      error: error.response?.data,
     })
   }
 }